Add POET and use cli to find orphans to remove
This needs to be revisited because there cases where it is not clear
if it is okay to remove certain files.

Right now, POET considers everything in the Interactive directory to be
orphaned.

POET also considers text files like
osbooks-principles-of-management-bundle/media/4-20-longdesc.txt to be
orphaned.

For now, we will just check the root directory of the book for orphans.

Also fixed a bug with tag and branch output parsing (it would return
[''], now it will not)